% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epictemplate_project.R
\name{create_epictemplate_project}
\alias{create_epictemplate_project}
\title{Create project}
\usage{
create_epictemplate_project(path, ...)
}
\arguments{
\item{path}{The path provided by the New Project Wizard}

\item{...}{Additional parameters: stan (checkbox)}
}
\value{
Creates a new project
}
\description{
Create project
}
