% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_file.R
\name{new_file}
\alias{new_file}
\title{Create new file}
\usage{
new_file()
}
\value{
Creates a new file according to the specifications.
}
\description{
Base function for the "Create file" addin. Inputs are provided via dialogs.
}
\examples{
\dontrun{
new_file()
}
}
